Use longjmp for runtime error management
Generating runtime errors from many levels down the callstack from
exec_func is costly. Functions have to return values in an error case
that might not make sense. Also we would have to check every iteration
if an error occured and exit the vm dispatch loop. Use longjmp to
circumvent this in a runtime error case.